validation function and green test
