+++ javascript-code-challenge/validation-test.js
@@ -1,14 +1,12 @@
-var regex = /(^[a-j]{1}$)|(^[NLQR]{1}$)|(^[NLQR](?=([a-j]|[ZNLQR]){2}))|(^[Z](?:_?[a-j]+)*$)/;
-var splitWith =/ +?/g;
-var validation = function(inputSting){
-    var string = inputSting.split(splitWith);
-    for(i=0; i < string.length; i++){
-        var valid = regex.test(string[i]);
-        if(valid){
-            console.log(string[i] + " VALID");
-        }else{
-            console.log(string[i] + " INVALID");
-        };
-    }
-};
-validation('Ze Za Nj QZja\nRhfa QRZZaZZaj ssfsfsf');
+var assert = require("assert");
+var fs = require('fs');
+var vm = require('vm');
+var path = './javascript-code-challenge/validation.js';
+var code = fs.readFileSync(path);
+vm.runInThisContext(code);
+
+describe('validation function', function(){
+    it('validates the input', function(){
+        assert.equal(validation('Ze'), 'Ze VALID');
+    })
+});
